+++ src/client/cypress.config.js
@@ -1,9 +1,18 @@
-const { defineConfig } = require("cypress")
+const { defineConfig } = require("cypress");
 
 module.exports = defineConfig({
   e2e: {
     baseUrl: "http://localhost:3000",
     video: false,
-    supportFile: false
-  }
-})
+    supportFile: false,
+    setupNodeEvents(on, config) {
+      on("task", {
+        log(message) {
+          console.log(message);
+
+          return null;
+        },
+      });
+    },
+  },
+});
